+++ src/ui/CircularKnob.h
@@ -3,7 +3,9 @@
 #include <juce_graphics/juce_graphics.h>
 #include <juce_gui_basics/juce_gui_basics.h>
 
-class CircularKnob : public juce::Slider
+#include "FineControlSlider.h"
+
+class CircularKnob : public FineControlSlider
 {
    public:
     bool hitTest(int mx, int my) override;